Add files via upload
A Mental illnesses are health condition involving changes in emotion, thinking or behavior (or a combination of these). 
Mental illnesses can be associated with distress and/or problems functioning in social, work or family activities.
Mental health is a huge issue in the United States and there are shocking statistics to prove it.